Split complex component in different subcomponents
